TSP-282 migrate telegramId from Integer to Long

+++ src/main/java/com/github/sportbot/dto/ExerciseEntryRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record ExerciseEntryRequest(
-        @NotNull Integer telegramId,
+        @NotNull Long telegramId,
         @NotBlank String exerciseType,
         int count
 ) {}
